Feat: add cookies message to flitepath.io
The flitepath.io (aka getflitepath.com) website should have a cookie usage policy that pops up when the user lands on the page.
The behavior when the user clicks [OK, I agree] should be the same as it is currently.
The behavior if the user does not click [OK, I agree] should be the same as it is currently.

Resolves: #1171900433772353